requirement.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-pinecone
pinecone-client
openai
